add social authentication and forget password option

+++ controllers/comments_controller.js
@@ -6,32 +6,36 @@ const commentsMailer = require('../mailers/comments_mailer');
 
 module.exports.create = async function (req, res) {
   try {
-    let post = await Post.findById(req.body.post);
-    if (!post) {
-      req.flash('error', 'Post not found while creating comment'); //console.log('Post not found while creating comment');
-      return res.redirect('back');
-    }
-    let comment = await Comments.create({
-      content: req.body.content,
-      user: req.user._id,
-      post: req.body.post,
-    });
-
-    post.comments.push(comment);
-    await post.save();
-    await comment.populate('user', 'name email');
-    commentsMailer.newComment(comment);
-
-    if(req.xhr){
-      return res.status(200).json({
-          data: {
-            comment: comment
-          },
-          message: "comment created"
+    if (req.isAuthenticated()){
+      let post = await Post.findById(req.body.post);
+      if (!post) {
+        req.flash('error', 'Post not found while creating comment'); //console.log('Post not found while creating comment');
+        return res.redirect('back');
+      }
+      let comment = await Comments.create({
+        content: req.body.content,
+        user: req.user._id,
+        post: req.body.post,
       });
+
+      post.comments.push(comment);
+      await post.save();
+      await comment.populate('user', 'name email');
+      commentsMailer.newComment(comment);
+
+      if(req.xhr){
+        return res.status(200).json({
+            data: {
+              comment: comment
+            },
+            message: "comment created"
+        });
+      }
+      req.flash('success', 'Comment updated successfully');
+      return res.redirect('/');
+    }else{
+      return res.redirect('/users/signin');
     }
-    req.flash('success', 'Comment updated successfully');
-    return res.redirect('/');
   } catch (err) {
     req.flash('error', err);
     return res.redirect('back');
@@ -40,23 +44,27 @@ module.exports.create = async function (req, res) {
 
 module.exports.destroy = async function(req, res){
   try{
-    let comment = await Comments.findById(req.params.id);
-    if (comment.user.toString() == req.user.id){
-      let postId = comment.post;
-      await Comments.deleteOne({ _id: req.params.id });
-      await Post.findByIdAndUpdate(postId, {$pull:{comments:req.params.id}});
-      await Like.deleteMany({likeable: comment._id, onModel: 'Comment'});
-      if(req.xhr){
-        return res.status(200).json({
-            data: {
-                comment_id: req.params.id
-            },
-            message: "comment deleted"
-        });
+    if (req.isAuthenticated()){
+      let comment = await Comments.findById(req.params.id);
+      if (comment.user.toString() == req.user.id){
+        let postId = comment.post;
+        await Comments.deleteOne({ _id: req.params.id });
+        await Post.findByIdAndUpdate(postId, {$pull:{comments:req.params.id}});
+        await Like.deleteMany({likeable: comment._id, onModel: 'Comment'});
+        if(req.xhr){
+          return res.status(200).json({
+              data: {
+                  comment_id: req.params.id
+              },
+              message: "comment deleted"
+          });
+        } 
+        req.flash('success', 'Comment deleted successfully');
       } 
-      req.flash('success', 'Comment deleted successfully');
-    } 
-    return res.redirect('back');
+      return res.redirect('back');
+    }else{
+      return res.redirect('/users/signin');
+    }
   } catch(err){
     req.flash('error', err);
     return res.redirect('back');
