+++ controllers/likes_controller.js
@@ -7,49 +7,53 @@ module.exports.toggleLike = async function(req, res){
     try{
         // console.log("inside like");
         // likes/toggle/?id=abcdef&type=Post
-        let likeable;
-        let deleted = false;
+        if (req.isAuthenticated()){
+            let likeable;
+            let deleted = false;
 
-        if (req.query.type == 'Post'){
-            likeable = await Post.findById(req.query.id).populate('likes');
-        }else{
-            likeable = await Comment.findById(req.query.id).populate('likes');
-        }
+            if (req.query.type == 'Post'){
+                likeable = await Post.findById(req.query.id).populate('likes');
+            }else{
+                likeable = await Comment.findById(req.query.id).populate('likes');
+            }
 
-        // check if a like already exists
-        let existingLike = await Like.findOne({
-            likeable: req.query.id,
-            onModel: req.query.type,
-            user: req.user.id
-        });
-
-        // if a like already exists then delete it
-        if (existingLike){
-            likeable.likes.pull(existingLike._id);
-            likeable.save();
-            await existingLike.deleteOne({
+            // check if a like already exists
+            let existingLike = await Like.findOne({
                 likeable: req.query.id,
                 onModel: req.query.type,
                 user: req.user.id
-            }); //delete post
-            // existingLike.remove();
-            deleted = true;
-        }else{
-            // else make a new like
-            let newLike = await Like.create({
-                user: req.user._id,
-                likeable: req.query.id,
-                onModel: req.query.type
             });
-            likeable.likes.push(newLike._id);
-            likeable.save();
-        }
-        return res.status(200).json({
-            message: "Request successful!",
-            data: {
-                deleted: deleted
+
+            // if a like already exists then delete it
+            if (existingLike){
+                likeable.likes.pull(existingLike._id);
+                likeable.save();
+                await existingLike.deleteOne({
+                    likeable: req.query.id,
+                    onModel: req.query.type,
+                    user: req.user.id
+                }); //delete post
+                // existingLike.remove();
+                deleted = true;
+            }else{
+                // else make a new like
+                let newLike = await Like.create({
+                    user: req.user._id,
+                    likeable: req.query.id,
+                    onModel: req.query.type
+                });
+                likeable.likes.push(newLike._id);
+                likeable.save();
             }
-        });
+            return res.status(200).json({
+                message: "Request successful!",
+                data: {
+                    deleted: deleted
+                }
+            });
+        }else{
+            return res.redirect('/users/signin');
+        }
     }catch(err){
         console.log(err);
         return res.status(500).json({
